Merge Cell Basic, Table, and Conditional Styles
Fix #1058, which went stale in 2019 and is now reopened. PR #4412 added support for Conditional Styles and partial support for Table Styles to Html Writer. This PR refactors that code by moving it to a new `Styles\ConditionalFormatting\MergedCellStyle` class, which will be used by Html Writer, and can be used in other circumstances. Note that Excel's builtin table styles don't contain the information we need to support them, but custom table styles are supported.